+++ qupath-gui-fx/src/main/java/qupath/lib/gui/viewer/ImageOverview.java
@@ -109,6 +109,9 @@ public class ImageOverview implements QuPathViewerListener {
 			e.consume();
 		});
 		
+		viewer.zPositionProperty().addListener(v -> repaint());
+		viewer.tPositionProperty().addListener(v -> repaint());
+			
 		viewer.addViewerListener(this);
 	}